+++ src/main/java/com/example/team1_be/domain/Day/Service/DayService.java
@@ -1,4 +1,4 @@
-package com.example.team1_be.domain.Day;
+package com.example.team1_be.domain.Day.Service;
 
 import java.util.List;
 
@@ -6,6 +6,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.example.team1_be.domain.Day.Day;
+import com.example.team1_be.domain.Day.DayRepository;
 import com.example.team1_be.domain.Week.Week;
 import com.example.team1_be.utils.errors.exception.CustomException;
